postgre-mcp/requirements.txt
psycopg2-binary
pandas
python-dotenv
fastapi
uvicorn
